+++ classes/privacy/provider.php
@@ -15,19 +15,19 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Theme Boost Union Child - Privacy provider
+ * Theme UR Courses - Privacy provider
  *
- * @package    theme_boost_union_child
+ * @package    theme_urcourses
  * @copyright  2023 Daniel Poggenpohl <daniel.poggenpohl@fernuni-hagen.de> and Alexander Bias <bias@alexanderbias.de>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace theme_boost_union_child\privacy;
+namespace theme_urcourses\privacy;
 
 /**
  * Privacy Subsystem implementing null_provider.
  *
- * @package    theme_boost_union_child
+ * @package    theme_urcourses
  * @copyright  2023 Daniel Poggenpohl <daniel.poggenpohl@fernuni-hagen.de> and Alexander Bias <bias@alexanderbias.de>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */