+++ settings.php
@@ -15,9 +15,9 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Theme Boost Union Child - Settings file
+ * Theme UR Courses - Settings file
  *
- * @package    theme_boost_union_child
+ * @package    theme_urcourses
  * @copyright  2023 Daniel Poggenpohl <daniel.poggenpohl@fernuni-hagen.de> and Alexander Bias <bias@alexanderbias.de>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
@@ -34,19 +34,19 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
 
     // However, there is still the $settings variable which is expected by Moodle core to be filled with the theme
     // settings and which is automatically linked from the theme selector page.
-    // To avoid that there appears a broken "Boost Union Child" settings page, we redirect the user to a settings
+    // To avoid that there appears a broken "UR Courses" settings page, we redirect the user to a settings
     // overview page if he opens this page.
-    $mainsettingspageurl = new moodle_url('/admin/settings.php', ['section' => 'themesettingboost_union_child']);
+    $mainsettingspageurl = new moodle_url('/admin/settings.php', ['section' => 'themesettingurcourses']);
     if ($ADMIN->fulltree && $PAGE->has_set_url() && $PAGE->url->compare($mainsettingspageurl)) {
-        redirect(new moodle_url('/admin/settings.php', ['section' => 'theme_boost_union_child']));
+        redirect(new moodle_url('/admin/settings.php', ['section' => 'theme_urcourses']));
     }
 
     // Create empty settings page structure to make the site administration work on non-admin pages.
     if (!$ADMIN->fulltree) {
-        // Create Boost Union Child settings page
+        // Create UR Courses settings page
         // (and allow users with the theme/boost_union:configure capability to access it).
-        $tab = new admin_settingpage('theme_boost_union_child',
-                get_string('configtitle', 'theme_boost_union_child', null, true),
+        $tab = new admin_settingpage('theme_urcourses',
+                get_string('configtitle', 'theme_urcourses', null, true),
                 'theme/boost_union:configure');
         $ADMIN->add('theme_boost_union', $tab);
     }
@@ -58,8 +58,8 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         // Require the necessary libraries.
         require_once($CFG->dirroot . '/theme/boost_union/lib.php');
         require_once($CFG->dirroot . '/theme/boost_union/locallib.php');
-        require_once($CFG->dirroot . '/theme/boost_union_child/lib.php');
-        require_once($CFG->dirroot . '/theme/boost_union_child/locallib.php');
+        require_once($CFG->dirroot . '/theme/urcourses/lib.php');
+        require_once($CFG->dirroot . '/theme/urcourses/locallib.php');
 
         // Prepare options array for select settings.
         // Due to MDL-58376, we will use binary select settings instead of checkbox settings throughout this theme.
@@ -67,48 +67,48 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
                 THEME_BOOST_UNION_SETTING_SELECT_NO => get_string('no'), ];
 
 
-        // Create Boost Union Child settings page with tabs
+        // Create UR Courses settings page with tabs
         // (and allow users with the theme/boost_union:configure capability to access it).
-        $page = new theme_boost_admin_settingspage_tabs('theme_boost_union_child',
-                get_string('configtitle', 'theme_boost_union_child', null, true),
+        $page = new theme_boost_admin_settingspage_tabs('theme_urcourses',
+                get_string('configtitle', 'theme_urcourses', null, true),
                 'theme/boost_union:configure');
 
 
         // Create general settings tab.
-        $tab = new admin_settingpage('theme_boost_union_child_general',
+        $tab = new admin_settingpage('theme_urcourses_general',
                 get_string('generalsettings', 'theme_boost', null, true));
 
         // Create inheritance heading.
-        $name = 'theme_boost_union_child/inheritanceheading';
-        $title = get_string('inheritanceheading', 'theme_boost_union_child', null, true);
+        $name = 'theme_urcourses/inheritanceheading';
+        $title = get_string('inheritanceheading', 'theme_urcourses', null, true);
         $setting = new admin_setting_heading($name, $title, null);
         $tab->add($setting);
 
         // Prepare inheritance options.
         $inheritanceoptions = [
-                THEME_BOOST_UNION_CHILD_SETTING_INHERITANCE_INHERIT =>
-                        get_string('inheritanceinherit', 'theme_boost_union_child'),
-                THEME_BOOST_UNION_CHILD_SETTING_INHERITANCE_DUPLICATE =>
-                        get_string('inheritanceduplicate', 'theme_boost_union_child'),
+                THEME_URCOURSES_SETTING_INHERITANCE_INHERIT =>
+                        get_string('inheritanceinherit', 'theme_urcourses'),
+                THEME_URCOURSES_SETTING_INHERITANCE_DUPLICATE =>
+                        get_string('inheritanceduplicate', 'theme_urcourses'),
         ];
 
         // Setting: Pre SCSS inheritance setting.
-        $name = 'theme_boost_union_child/prescssinheritance';
-        $title = get_string('prescssinheritancesetting', 'theme_boost_union_child', null, true);
-        $description = get_string('prescssinheritancesetting_desc', 'theme_boost_union_child', null, true).'<br />'.
-                get_string('inheritanceoptionsexplanation', 'theme_boost_union_child', null, true);
+        $name = 'theme_urcourses/prescssinheritance';
+        $title = get_string('prescssinheritancesetting', 'theme_urcourses', null, true);
+        $description = get_string('prescssinheritancesetting_desc', 'theme_urcourses', null, true).'<br />'.
+                get_string('inheritanceoptionsexplanation', 'theme_urcourses', null, true);
         $setting = new admin_setting_configselect($name, $title, $description,
-                THEME_BOOST_UNION_CHILD_SETTING_INHERITANCE_INHERIT, $inheritanceoptions);
+                THEME_URCOURSES_SETTING_INHERITANCE_INHERIT, $inheritanceoptions);
         $setting->set_updatedcallback('theme_reset_all_caches');
         $tab->add($setting);
 
         // Setting: Extra SCSS inheritance setting.
-        $name = 'theme_boost_union_child/extrascssinheritance';
-        $title = get_string('extrascssinheritancesetting', 'theme_boost_union_child', null, true);
-        $description = get_string('extrascssinheritancesetting_desc', 'theme_boost_union_child', null, true).'<br />'.
-                get_string('inheritanceoptionsexplanation', 'theme_boost_union_child', null, true);
+        $name = 'theme_urcourses/extrascssinheritance';
+        $title = get_string('extrascssinheritancesetting', 'theme_urcourses', null, true);
+        $description = get_string('extrascssinheritancesetting_desc', 'theme_urcourses', null, true).'<br />'.
+                get_string('inheritanceoptionsexplanation', 'theme_urcourses', null, true);
         $setting = new admin_setting_configselect($name, $title, $description,
-                THEME_BOOST_UNION_CHILD_SETTING_INHERITANCE_INHERIT, $inheritanceoptions);
+                THEME_URCOURSES_SETTING_INHERITANCE_INHERIT, $inheritanceoptions);
         $setting->set_updatedcallback('theme_reset_all_caches');
         $tab->add($setting);
 
@@ -117,7 +117,7 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
 
         /**********************************************************
          * EXTENSION POINT:
-         * Add your Boost Union Child settings here.
+         * Add your UR Courses settings here.
          *********************************************************/
 
         // Add settings page to the admin settings category.