+++ tests/behat/behat_theme_urcourses_behat_course.php
@@ -15,9 +15,9 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Behat grade-related step definition overrides for the Boost Union Child theme.
+ * Behat course-related step definition overrides for the UR Courses theme.
  *
- * @package    theme_boost_union_child
+ * @package    theme_urcourses
  * @category   test
  * @copyright  2022 Luca Bösch, BFH Bern University of Applied Sciences luca.boesch@bfh.ch
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
@@ -26,17 +26,20 @@
 // NOTE: no MOODLE_INTERNAL test here, this file may be required by behat before including /config.php.
 // For that reason, we can't even rely on $CFG->admin being available here.
 
-require_once(__DIR__ . '/../../../boost_union/tests/behat/behat_theme_boost_union_behat_grade.php');
+require_once(__DIR__ . '/../../../boost_union/tests/behat/behat_theme_boost_union_behat_course.php');
 
-use Behat\Gherkin\Node\TableNode as TableNode;
+use Behat\Gherkin\Node\TableNode as TableNode,
+    Behat\Mink\Exception\ExpectationException as ExpectationException,
+    Behat\Mink\Exception\DriverException as DriverException,
+    Behat\Mink\Exception\ElementNotFoundException as ElementNotFoundException;
 
 /**
- * Grade-related step definition overrides for the Boost Union Child theme.
+ * Course-related step definition overrides for the UR Courses theme.
  *
- * @package    theme_boost_union_child
+ * @package    theme_urcourses
  * @category   test
  * @copyright  2022 Luca Bösch, BFH Bern University of Applied Sciences luca.boesch@bfh.ch
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class behat_theme_boost_union_child_behat_grade extends behat_theme_boost_union_behat_grade {
+class behat_theme_urcourses_behat_course extends behat_theme_boost_union_behat_course {
 }