+++ CHANGES.md
@@ -1,4 +1,4 @@
-moodle-theme_boost_union_child
+moodle-theme_urcourses
 ==============================
 
 Changes
@@ -10,7 +10,7 @@ Changes
 * 2024-07-25 - Development: Rename master branch to main, please update your clones.
 * 2024-07-23 - Change Moodle-Plugin-CI to run on Moodle Core 4.4 and Boost Union 4.4
 * 2024-07-23 - Make codechecker happy again.
-* 2024-04-06 - Adopt Boost Union Child settings to the settings changes in the upcoming Boost Union v4.3-r11 release and for Moodle 4.4.
+* 2024-04-06 - Adopt UR Courses settings to the settings changes in the upcoming Boost Union v4.3-r11 release and for Moodle 4.4.
 * 2024-03-13 - Make sure that all Boost Union settings are added to the $THEME object even if the Boost Union config was already included in some other place, resolves #3.
 * 2024-02-23 - Add example branch which shows how to modify a mustache template from Moodle core.
 * 2024-02-23 - Add example branch which shows how to allow the admin to configure a SCSS variable with an admin setting.